latex works and an updated 01 and builkd.py

+++ papers/build.py
@@ -13,6 +13,7 @@ from subprocess import run
 import webbrowser as wb
 from pathlib import Path
 from shutil import copy, copytree
+import yaml 
  
 import sys
 print(sys.argv)
@@ -28,12 +29,12 @@ else:
 doall = '--all' if 'all' in options else ''
 buildloc = Path(f'{bookdir}/_build/')
 buildhtml = buildloc / 'html'
-# (destination := Path(fr'C:/modelbook/IbHansen.github.io/{bookdir}')).mkdir(parents=True, exist_ok=True)
 fileloc = str((buildhtml / 'index.html').absolute())
 
-#print(f'{fileloc=}\n{destination=}')
-print(f'{fileloc=}\n') #dropped destination
-# breakpoint()
+print(f'{fileloc=}\n')
+
+    
+assert 1==1
 xx0 = run(f'jb build {bookdir}/ {doall}')
 # wb.open(rf'file://C:\wb new\Modelflow\working_paper\{bookdir}\_build\html\index.html', new=2)
 wb.open(rf'file://{fileloc}', new=2)
@@ -65,10 +66,24 @@ for dir in sorted(Path(f'{bookdir}/_build/jupyter_execute').glob('**')):
             # print(f'Not copied{picture}')
      
 if 'latex-pdf' in options: 
-     xx0 = run(f'jb build {bookdir}/ --builder=latex')
-     xx0 = run('latexmk -pdf -dvi- -ps- -f MFModinModelflow.tex',cwd = f'{bookdir}/_build/latex/')
+    #%% Find the name of the .tex file. 
+    with open(Path(f'{bookdir}/_config.yml')) as f: 
+        config_dict = yaml.safe_load(f)
+    try:    
+        texfile = config_dict['latex']['latex_documents']['targetname']
+    except: 
+        texfile = 'book.tex'    
+    #%% 
+    xx0 = run(f'jb build {bookdir}/ --builder=latex')
+    xx0 = run(f'latexmk -pdf -dvi- -ps- -f {texfile}',cwd = f'{bookdir}/_build/latex/')
      
 if 'copy' in options:
-    copytree(buildhtml,destination,dirs_exist_ok=True )
+    checkdir = Path(fr'C:/modelbook/IbHansen.github.io')
+    if checkdir.exists():
+        (destination := Path(fr'C:/modelbook/IbHansen.github.io/{bookdir}')).mkdir(parents=True, exist_ok=True)
+        print(f'Hope you are Ib, Now copy the html files to: {destination=}\n')
+        copytree(buildhtml,destination,dirs_exist_ok=True )
+    else:
+        print('sorry you seem not to be Ib so yu cant copy html files to local github location')
